backend/requirements.txt
django
python-decouple
psycopg[binary]


channels[daphne]
channels_redis


selenium